requirements.txt
# grovepi requirements
grovepi

# Cryptography
cryptography==3.3.2

# GCP requirements
cloud-sql-python-connector[pymysql]
sqlalchemy
python-dotenv

# User Interface
flask
turbo-flask

# MQTT
paho-mqtt